+++ client/src/products/about/keyword-input/ComboboxProvider.tsx
@@ -1,12 +1,11 @@
 import { forwardRef } from "react";
-import { FilteredOptionsProvider } from "./FilteredOptions/filteredOptionsContext";
 import { InputContextProvider } from "felleskomponenter/comboboxfelles/Input/Input.context";
 import { SelectedOptionsProvider } from "felleskomponenter/comboboxfelles/SelectedOptions/selectedOptionsContext";
 import { mapToComboboxOptionArray } from "felleskomponenter/comboboxfelles/combobox-utils";
-import { IsoBoxProps } from "./index";
-import IsoCombobox from "products/iso-combobox/combobox/IsoCombobox";
+import { ComboboxProps } from "felleskomponenter/comboboxfelles/types";
+import { KeywordCombobox } from "products/about/keyword-input/KeywordCombobox";
 
-const ComboboxProvider = forwardRef<HTMLInputElement, IsoBoxProps>(function ComboboxProvider(props, ref) {
+const ComboboxProvider = forwardRef<HTMLInputElement, ComboboxProps>(function ComboboxProvider(props, ref) {
   const {
     children,
     error,
@@ -18,14 +17,10 @@ const ComboboxProvider = forwardRef<HTMLInputElement, IsoBoxProps>(function Comb
     value,
     onChange,
     onClear,
-    filteredOptions: externalFilteredOptions,
-    isListOpen,
-    isLoading = false,
     ...rest
   } = props;
   const options = mapToComboboxOptionArray(externalOptions) || [];
   const selectedOptions = mapToComboboxOptionArray(externalSelectedOptions);
-  const filteredOptions = mapToComboboxOptionArray(externalFilteredOptions);
   return (
     <InputContextProvider
       value={{
@@ -40,23 +35,14 @@ const ComboboxProvider = forwardRef<HTMLInputElement, IsoBoxProps>(function Comb
       <SelectedOptionsProvider
         value={{
           selectedOptions,
+          maxSelected,
           onToggleSelected,
           options,
-          maxSelected,
         }}
       >
-        <FilteredOptionsProvider
-          value={{
-            filteredOptions,
-            isListOpen,
-            isLoading,
-            options,
-          }}
-        >
-          <IsoCombobox ref={ref} {...rest}>
-            {children}
-          </IsoCombobox>
-        </FilteredOptionsProvider>
+        <KeywordCombobox ref={ref} {...rest}>
+          {children}
+        </KeywordCombobox>
       </SelectedOptionsProvider>
     </InputContextProvider>
   );
