+++ client/src/products/iso-combobox/FilteredOptions/filtered-options-util.ts
@@ -1,4 +1,4 @@
-import { ComboboxOption } from "../index";
+import { ComboboxOption } from "felleskomponenter/comboboxfelles/types";
 
 const normalizeText = (text: string): string => (typeof text === "string" ? text.toLocaleLowerCase().trim() : "");
