+++ client/src/products/iso-combobox/FilteredOptions/filteredOptionsContext.tsx
@@ -3,7 +3,8 @@ import React, { SetStateAction, useCallback, useMemo, useState } from "react";
 import { createContext } from "felleskomponenter/comboboxfelles/utils/create-context";
 import { useClientLayoutEffect, usePrevious } from "felleskomponenter/comboboxfelles/utils";
 import { useInputContext } from "felleskomponenter/comboboxfelles/Input/Input.context";
-import { ComboboxOption, IsoBoxProps } from "../index";
+import { ComboboxOption } from "felleskomponenter/comboboxfelles/types";
+import { IsoBoxProps } from "../types";
 import filteredOptionsUtils from "./filtered-options-util";
 import useVirtualFocus, { VirtualFocusType } from "./useVirtualFocus";
 import { useSelectedOptionsContext } from "felleskomponenter/comboboxfelles/SelectedOptions/selectedOptionsContext";
