+++ client/src/products/iso-combobox/index.ts
@@ -1,2 +1,1 @@
 export { default as ISO_Combobox } from "./ComboboxProvider";
-export { type IsoBoxProps } from "./types";
